successfully renders group
